+++ examples/simpletri2.cpp
@@ -14,20 +14,23 @@ static void simple_tri_2_engine_keyboard_callback(GLFWwindow *window, int key, i
 class SimpleTri2Engine final : public CSCI441::OpenGLEngine {
 public:
     SimpleTri2Engine(
-        int OPENGL_MAJOR_VERSION, int OPENGL_MINOR_VERSION,
-        int WINDOW_WIDTH, int WINDOW_HEIGHT,
+        const int OPENGL_MAJOR_VERSION, const int OPENGL_MINOR_VERSION,
+        const int WINDOW_WIDTH, const int WINDOW_HEIGHT,
         const char* WINDOW_TITLE
     ) : CSCI441::OpenGLEngine(OPENGL_MAJOR_VERSION, OPENGL_MINOR_VERSION, WINDOW_WIDTH, WINDOW_HEIGHT, WINDOW_TITLE),
         _triangleVAO(0),
         _triforceRotationSpeed(1.0f),
-        _triforceAngle(0.0f) { }
+        _triforceAngle(0.0f)
+    {
 
-    void run() final {
+    }
+
+    void run() override {
         // update the projection matrix based on the window size
         // the GL_PROJECTION matrix governs properties of the view coordinates;
         // i.e. what gets seen - use an Orthographic projection that ranges
         // from [0, windowWidth] in X and [0, windowHeight] in Y. (0,0) is the lower left.
-        glm::mat4 projMtx = glm::ortho( 0.0f, (GLfloat)mWindowWidth, 0.0f, (GLfloat)mWindowHeight );
+        const glm::mat4 projMtx = glm::ortho( 0.0f, static_cast<GLfloat>(mWindowWidth), 0.0f, static_cast<GLfloat>(mWindowHeight) );
         CSCI441::SimpleShader2::setProjectionMatrix(projMtx);
 
         // Get the size of our framebuffer.  Ideally this should be the same dimensions as our window, but
@@ -46,9 +49,12 @@ public:
         //  window will display once and then the program exits.
         while( !glfwWindowShouldClose(mpWindow) ) { // check if the window was instructed to be closed
             glClear( GL_COLOR_BUFFER_BIT );           // clear the current color contents in the window
+
             _renderScene();                                 // draw everything to the window
-            glfwSwapBuffers(mpWindow);              // flush the OpenGL commands and make sure they get rendered!
+
+            glfwSwapBuffers(mpWindow);                      // flush the OpenGL commands and make sure they get rendered!
             glfwPollEvents();                               // check for any events and signal to redraw screen
+
             currTime = glfwGetTime();
             _updateScene(lastTime - currTime);          // animate scene components
             lastTime = currTime;
@@ -56,17 +62,17 @@ public:
     }
 
 protected:
-    void mSetupGLFW() final {
+    void mSetupGLFW() override {
         CSCI441::OpenGLEngine::mSetupGLFW();
         glfwSetKeyCallback( mpWindow, simple_tri_2_engine_keyboard_callback );
     }
-    void mSetupOpenGL() final {
+    void mSetupOpenGL() override {
         glClearColor( 0.0f, 0.0f, 0.0f, 1.0f );     // set the clear color to black
     }
-    void mSetupShaders() final {
+    void mSetupShaders() override {
         CSCI441::SimpleShader2::setupSimpleShader();
     }
-    void mSetupBuffers() final {
+    void mSetupBuffers() override {
         glm::vec3 gold(0.9f, 0.8f, 0.1f);
 
         _trianglePoints.emplace_back( -2.5f, -2.0f );   _triangleColorsGold.emplace_back( gold );
@@ -75,14 +81,21 @@ protected:
 
         _triangleVAO = CSCI441::SimpleShader2::registerVertexArray( _trianglePoints, _triangleColorsGold );
     }
+    void mCleanupBuffers() override {
+        CSCI441::SimpleShader2::deleteVertexArray( _triangleVAO );
+        _triangleVAO = 0;
+
+        _trianglePoints.clear();
+        _triangleColorsGold.clear();
+    }
 
 private:
     void _renderScene() const {
-        glm::mat4 transMtx = glm::translate( glm::mat4(1.0), glm::vec3(getWindowWidth() / 2.0f, getWindowHeight() / 2.0f, 0.0f) );
+        const glm::mat4 transMtx = glm::translate( glm::mat4(1.0), glm::vec3(getWindowWidth() / 2.0f, getWindowHeight() / 2.0f, 0.0f) );
         CSCI441::SimpleShader2::pushTransformation( transMtx );
-            glm::mat4 rotMtx = glm::rotate( glm::mat4(1.0), _triforceAngle, glm::vec3(0, 0, 1) );
+            const glm::mat4 rotMtx = glm::rotate( glm::mat4(1.0), _triforceAngle, glm::vec3(0, 0, 1) );
             CSCI441::SimpleShader2::pushTransformation( rotMtx );
-                glm::mat4 scaleMtx = glm::scale( glm::mat4(1.0), glm::vec3(10, 10, 1) );
+                const glm::mat4 scaleMtx = glm::scale( glm::mat4(1.0), glm::vec3(10, 10, 1) );
                 CSCI441::SimpleShader2::pushTransformation( scaleMtx );
                     _drawTriforce();
                 CSCI441::SimpleShader2::popTransformation();
@@ -95,25 +108,25 @@ private:
     }
 
     void _drawTriforce() const {
-        glm::mat4 t1 = glm::translate( glm::mat4(1.0), glm::vec3( -2.5f, -2.0f, 0.0f ) );
+        constexpr glm::mat4 t1 = glm::translate( glm::mat4(1.0), glm::vec3( -2.5f, -2.0f, 0.0f ) );
         CSCI441::SimpleShader2::pushTransformation( t1 );
             _drawTriangle();
         CSCI441::SimpleShader2::popTransformation();
 
-        glm::mat4 t2 = glm::translate( glm::mat4(1.0), glm::vec3( 2.5f, -2.0f, 0.0f ) );
+        constexpr glm::mat4 t2 = glm::translate( glm::mat4(1.0), glm::vec3( 2.5f, -2.0f, 0.0f ) );
         CSCI441::SimpleShader2::pushTransformation( t2 );
             _drawTriangle();
         CSCI441::SimpleShader2::popTransformation();
 
-        glm::mat4 t3 = glm::translate( glm::mat4(1.0), glm::vec3( 0.0f, 2.0f, 0.0f ) );
+        constexpr glm::mat4 t3 = glm::translate( glm::mat4(1.0), glm::vec3( 0.0f, 2.0f, 0.0f ) );
         CSCI441::SimpleShader2::pushTransformation( t3 );
             _drawTriangle();
         CSCI441::SimpleShader2::popTransformation();
     }
 
-    void _updateScene(GLfloat dt) {
+    void _updateScene(const GLfloat dt) {
         _triforceAngle += _triforceRotationSpeed * dt;
-        if( _triforceAngle > 6.28f ) _triforceAngle -= 6.28f;
+        if( _triforceAngle > glm::two_pi<GLfloat>() ) _triforceAngle -= glm::two_pi<GLfloat>();
     }
 
     GLuint _triangleVAO;
@@ -123,8 +136,8 @@ private:
     GLfloat _triforceAngle;
 };
 
-void simple_tri_2_engine_keyboard_callback(GLFWwindow *window, int key, int scancode, int action, int mods) {
-    auto engine = (SimpleTri2Engine*) glfwGetWindowUserPointer(window);
+void simple_tri_2_engine_keyboard_callback(GLFWwindow *window, const int key, const int scancode, const int action, const int mods) {
+    const auto engine = static_cast<SimpleTri2Engine *>(glfwGetWindowUserPointer(window));
     if(action == GLFW_PRESS) {
         switch(key) {
             case GLFW_KEY_ESCAPE:
@@ -139,7 +152,7 @@ void simple_tri_2_engine_keyboard_callback(GLFWwindow *window, int key, int scan
 }
 
 int main() {
-    auto pSimpleTri2Engine = new SimpleTri2Engine(4, 1, 512, 512, "SimpleShader2 Triangle");
+    const auto pSimpleTri2Engine = new SimpleTri2Engine(4, 1, 512, 512, "SimpleShader2 Triangle");
     pSimpleTri2Engine->initialize();
     if (pSimpleTri2Engine->getError() == CSCI441::OpenGLEngine::OPENGL_ENGINE_ERROR_NO_ERROR) {
         pSimpleTri2Engine->run();